cloud_function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
pymongo==4.*
python-dotenv==1.*
